1. 配置和脚本增加了内置属性，内置的数据不允许删除，只可以修改

+++ code-factory/src/main/java/me/zhengjie/codefactory/service/impl/ConfigServiceImpl.java
@@ -100,7 +100,10 @@ public class ConfigServiceImpl implements ConfigService {
     @Override
     public void deleteAll(Long[] ids) {
         for (Long id : ids) {
-            configRepository.deleteById(id);
+            final Config config = configRepository.getById(id);
+            if(config!= null && !config.getBuidIn()){
+                configRepository.deleteById(id);
+            }
         }
     }
